# feat(integer): improve scalar shifts performances