--- a/tfhe/src/integer/server_key/radix_parallel/tests.rs
+++ b/tfhe/src/integer/server_key/radix_parallel/tests.rs
@@ -831,31 +831,43 @@ fn integer_unchecked_scalar_left_shift(param: Parameters) {
     let (cks, sks) = KEY_CACHE.get_from_params(param);
     let cks = RadixClientKey::from((cks, NB_CTXT));
 
-    //RNG
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
     let modulus = param.message_modulus.0.pow(NB_CTXT as u32) as u64;
-
-    //Nb of bits to shift
-    let tmp_f64 = param.message_modulus.0 as f64;
-    let nb_bits = tmp_f64.log2().floor() as usize * NB_CTXT;
+    let nb_bits = modulus.ilog2();
 
     for _ in 0..NB_TEST {
         let clear = rng.gen::<u64>() % modulus;
+        let scalar = rng.gen::<u32>();
 
-        let scalar = rng.gen::<usize>() % nb_bits;
-
-        // encryption of an integer
         let ct = cks.encrypt(clear);
 
-        let ct_res = sks.unchecked_scalar_left_shift_parallelized(&ct, scalar);
+        // case when 0<= scalar < nb_bits
+        {
+            let scalar = scalar % nb_bits;
+            let ct_res = sks.unchecked_scalar_left_shift_parallelized(&ct, scalar as u64);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shl(scalar).unwrap_or(0) % modulus, dec_res);
+        }
 
-        // decryption of ct_res
+        // case when scalar >= nb_bits
+        {
+            let scalar = scalar.saturating_add(nb_bits);
+            let ct_res = sks.unchecked_scalar_left_shift_parallelized(&ct, scalar as u64);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shl(scalar).unwrap_or(0) % modulus, dec_res);
+        }
+    }
+
+    let clear = rng.gen::<u64>() % modulus;
+    let ct = cks.encrypt(clear);
+
+    let nb_bits_in_block = param.message_modulus.0.ilog2();
+    for scalar in 0..nb_bits_in_block {
+        let ct_res = sks.unchecked_scalar_left_shift_parallelized(&ct, scalar as u64);
         let dec_res: u64 = cks.decrypt(&ct_res);
-
-        // assert
-        assert_eq!((clear << scalar) % modulus, dec_res);
+        assert_eq!(clear.checked_shl(scalar).unwrap_or(0) % modulus, dec_res);
     }
 }
 
@@ -863,34 +875,48 @@ fn integer_default_scalar_left_shift(param: Parameters) {
     let (cks, sks) = KEY_CACHE.get_from_params(param);
     let cks = RadixClientKey::from((cks, NB_CTXT));
 
-    //RNG
     let mut rng = rand::thread_rng();
 
-    // message_modulus^vec_length
     let modulus = param.message_modulus.0.pow(NB_CTXT as u32) as u64;
-
-    //Nb of bits to shift
-    let tmp_f64 = param.message_modulus.0 as f64;
-    let nb_bits = tmp_f64.log2().floor() as usize * NB_CTXT;
+    let nb_bits = modulus.ilog2();
 
     for _ in 0..NB_TEST {
         let clear = rng.gen::<u64>() % modulus;
+        let scalar = rng.gen::<u32>();
 
-        let scalar = rng.gen::<usize>() % nb_bits;
-
-        // encryption of an integer
         let ct = cks.encrypt(clear);
 
-        let ct_res = sks.scalar_left_shift_parallelized(&ct, scalar);
-        let tmp = sks.scalar_left_shift_parallelized(&ct, scalar);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp);
+        // case when 0<= scalar < nb_bits
+        {
+            let scalar = scalar % nb_bits;
+            let ct_res = sks.scalar_left_shift_parallelized(&ct, scalar as u64);
+            let tmp = sks.scalar_left_shift_parallelized(&ct, scalar as u64);
+            assert!(ct_res.block_carries_are_empty());
+            assert_eq!(ct_res, tmp);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shl(scalar).unwrap_or(0) % modulus, dec_res);
+        }
 
-        // decryption of ct_res
+        // case when scalar >= nb_bits
+        {
+            let scalar = scalar.saturating_add(nb_bits);
+            let ct_res = sks.scalar_left_shift_parallelized(&ct, scalar as u64);
+            let tmp = sks.scalar_left_shift_parallelized(&ct, scalar as u64);
+            assert!(ct_res.block_carries_are_empty());
+            assert_eq!(ct_res, tmp);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shl(scalar).unwrap_or(0) % modulus, dec_res);
+        }
+    }
+
+    let clear = rng.gen::<u64>() % modulus;
+    let ct = cks.encrypt(clear);
+
+    let nb_bits_in_block = param.message_modulus.0.ilog2();
+    for scalar in 0..nb_bits_in_block {
+        let ct_res = sks.scalar_left_shift_parallelized(&ct, scalar as u64);
         let dec_res: u64 = cks.decrypt(&ct_res);
-
-        // assert
-        assert_eq!((clear << scalar) % modulus, dec_res);
+        assert_eq!(clear.checked_shl(scalar).unwrap_or(0) % modulus, dec_res);
     }
 }
 
@@ -898,31 +924,52 @@ fn integer_unchecked_scalar_right_shift(param: Parameters) {
     let (cks, sks) = KEY_CACHE.get_from_params(param);
     let cks = RadixClientKey::from((cks, NB_CTXT));
 
-    //RNG
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
     let modulus = param.message_modulus.0.pow(NB_CTXT as u32) as u64;
-
-    //Nb of bits to shift
-    let tmp_f64 = param.message_modulus.0 as f64;
-    let nb_bits = tmp_f64.log2().floor() as usize * NB_CTXT;
+    let nb_bits = modulus.ilog2();
 
     for _ in 0..NB_TEST {
         let clear = rng.gen::<u64>() % modulus;
+        let scalar = rng.gen::<u32>();
 
-        let scalar = rng.gen::<usize>() % nb_bits;
-
-        // encryption of an integer
         let ct = cks.encrypt(clear);
 
-        let ct_res = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar);
+        // case when 0<= scalar < nb_bits
+        {
+            let scalar = scalar % nb_bits;
+            let ct_res = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar as u64);
+            let tmp = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar as u64);
+            assert!(ct_res.block_carries_are_empty());
+            assert_eq!(ct_res, tmp);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shr(scalar).unwrap_or(0) % modulus, dec_res);
+        }
 
-        // decryption of ct_res
+        // case when scalar >= nb_bits
+        {
+            let scalar = scalar.saturating_add(nb_bits);
+            let ct_res = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar as u64);
+            let tmp = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar as u64);
+            assert!(ct_res.block_carries_are_empty());
+            assert_eq!(ct_res, tmp);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shr(scalar).unwrap_or(0) % modulus, dec_res);
+        }
+    }
+
+    let clear = rng.gen::<u64>() % modulus;
+
+    let ct = cks.encrypt(clear);
+    let nb_bits_in_block = param.message_modulus.0.ilog2();
+    for scalar in 0..nb_bits_in_block {
+        let ct_res = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar as u64);
+        let tmp = sks.unchecked_scalar_right_shift_parallelized(&ct, scalar as u64);
+        assert!(ct_res.block_carries_are_empty());
+        assert_eq!(ct_res, tmp);
         let dec_res: u64 = cks.decrypt(&ct_res);
-
-        // assert
-        assert_eq!(clear >> scalar, dec_res);
+        assert_eq!(clear.checked_shr(scalar).unwrap_or(0) % modulus, dec_res);
     }
 }
 
@@ -930,34 +977,52 @@ fn integer_default_scalar_right_shift(param: Parameters) {
     let (cks, sks) = KEY_CACHE.get_from_params(param);
     let cks = RadixClientKey::from((cks, NB_CTXT));
 
-    //RNG
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
     let modulus = param.message_modulus.0.pow(NB_CTXT as u32) as u64;
-
-    //Nb of bits to shift
-    let tmp_f64 = param.message_modulus.0 as f64;
-    let nb_bits = tmp_f64.log2().floor() as usize * NB_CTXT;
+    let nb_bits = modulus.ilog2();
 
     for _ in 0..NB_TEST {
         let clear = rng.gen::<u64>() % modulus;
+        let scalar = rng.gen::<u32>();
 
-        let scalar = rng.gen::<usize>() % nb_bits;
-
-        // encryption of an integer
         let ct = cks.encrypt(clear);
 
-        let ct_res = sks.scalar_right_shift_parallelized(&ct, scalar);
-        let tmp = sks.scalar_right_shift_parallelized(&ct, scalar);
+        // case when 0<= scalar < nb_bits
+        {
+            let scalar = scalar % nb_bits;
+            let ct_res = sks.scalar_right_shift_parallelized(&ct, scalar as u64);
+            let tmp = sks.scalar_right_shift_parallelized(&ct, scalar as u64);
+            assert!(ct_res.block_carries_are_empty());
+            assert_eq!(ct_res, tmp);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shr(scalar).unwrap_or(0) % modulus, dec_res);
+        }
+
+        // case when scalar >= nb_bits
+        {
+            let scalar = scalar.saturating_add(nb_bits);
+            let ct_res = sks.scalar_right_shift_parallelized(&ct, scalar as u64);
+            let tmp = sks.scalar_right_shift_parallelized(&ct, scalar as u64);
+            assert!(ct_res.block_carries_are_empty());
+            assert_eq!(ct_res, tmp);
+            let dec_res: u64 = cks.decrypt(&ct_res);
+            assert_eq!(clear.checked_shr(scalar).unwrap_or(0) % modulus, dec_res);
+        }
+    }
+
+    let clear = rng.gen::<u64>() % modulus;
+
+    let ct = cks.encrypt(clear);
+    let nb_bits_in_block = param.message_modulus.0.ilog2();
+    for scalar in 0..nb_bits_in_block {
+        let ct_res = sks.scalar_right_shift_parallelized(&ct, scalar as u64);
+        let tmp = sks.scalar_right_shift_parallelized(&ct, scalar as u64);
         assert!(ct_res.block_carries_are_empty());
         assert_eq!(ct_res, tmp);
-
-        // decryption of ct_res
         let dec_res: u64 = cks.decrypt(&ct_res);
-
-        // assert
-        assert_eq!(clear >> scalar, dec_res);
+        assert_eq!(clear.checked_shr(scalar).unwrap_or(0) % modulus, dec_res);
     }
 }