@tmontaigu 2023-04-26
@slab-ci cpu_test